Upped max tables in effect

+++ spanner-data-validator-java/src/main/java/com/google/migration/common/DVTOptionsCore.java
@@ -276,7 +276,7 @@ public interface DVTOptionsCore extends PipelineOptions {
    * Max tables in effect at one time
    */
   @Description("Max tables in effect at one time")
-  @Default.Integer(100)
+  @Default.Integer(1000)
   Integer getMaxTablesInEffectAtOneTime();
 
   void setMaxTablesInEffectAtOneTime(Integer value);
